Clean up line lengths
